deleted leftover HomePage file in components
